Permissions

+++ backend/src/resolvers/Mutation.js
@@ -3,14 +3,22 @@ const jwt = require('jsonwebtoken');
 const { randomBytes } = require('crypto');
 const { promisify } = require('util');
 const { transport, makeANiceEmail } = require('../mail');
+const { hasPermission } = require('../utils');
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
-    // ToDo: Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
     const item = await ctx.db.mutation.createItem(
       {
         data: {
           ...args,
+          user: {
+            connect: {
+              id: ctx.request.userId,
+            },
+          },
         },
       },
       info
@@ -20,7 +28,9 @@ const Mutations = {
   },
 
   async updateItem(parent, args, ctx, info) {
-    // ToDo: Check if they are logged in
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
     const updates = { ...args };
     delete updates.id;
 
@@ -38,8 +48,18 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
     // find the item
-    const item = await ctx.db.query.item({ where }, info, `{ id, title }`);
+    const item = await ctx.db.query.item(
+      { where },
+      `{ id, title, user { id }}`
+    );
     // check ownership of item/permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permissions to delete this item!");
+    }
     // delete it
     return ctx.db.mutation.deleteItem({ where }, info);
   },
@@ -159,6 +179,29 @@ const Mutations = {
     });
 
     return updatedUser;
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // User logged in?
+    if (!ctx.request.userId) {
+      throw new Error('You must be logged in to do that!');
+    }
+    // Has permissions to update permissions
+    hasPermission(ctx.request.user, ['ADMIN', 'PERMISSIONUPDATE']);
+
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: {
+          id: args.userId,
+        },
+      },
+      info
+    );
   },
 };
